add authorization rules for task statuses actions

+++ resources/views/task_status/index.blade.php
@@ -1,10 +1,11 @@
 @extends('layouts.app')
 
 @section('content')
-<div class="container">
     <h1 class="mb-5">Статусы</h1>
 
-    <a href="{{ route('task_statuses.create') }}" class="btn btn-primary">Создать статус</a>
+    @can('create')
+        <a href="{{ route('task_statuses.create') }}" class="btn btn-primary">Создать статус</a>
+    @endcan
 
     <table class="table mt-2">
         <thead>
@@ -12,7 +13,9 @@
                 <th>ID</th>
                 <th>Имя</th>
                 <th>Дата создания</th>
-                <th>Действия</th>
+                @can('destroy')
+                    <th>Действия</th>
+                @endcan
             </tr>
         </thead>
         <tbody>
@@ -21,14 +24,15 @@
                 <td>{{ $status->id }}</td>
                 <td>{{ $status->name }}</td>
                 <td>{{ $status->created_at }}</td>
-                <td>
-                    <a class="text-danger" href="" data-confirm="Вы уверены?" data-method="delete">Удалить</a>
-                    <a href="">Изменить</a>
-                </td>
+                @can('destroy')
+                    <td>
+                        <a class="text-danger" href="" data-confirm="Вы уверены?" data-method="delete">Удалить</a>
+                        <a href="{{ route('task_statuses.edit', $status) }}">Изменить</a>
+                    </td>
+                @endcan
             </tr>
             @endforeach
         </tbody>
     </table>
     {{ $taskStatuses->links() }}
-</div>
 @endsection
